tmux: you don't need to be a login shell.
Running a login shell on macOS results in the path helper being run,
which results in the system paths being moved to the beginning of $PATH,
which results in them being earlier than asdf's shims directory.  Since
asdf's shims directory is already in the $PATH from the *original*
login shell, asdf doesn't re-add it, and it stays in the wrong place.